requirements.txt
Django==2.0.7
django-cors-headers==2.4.0
django-filter==2.0.0
djangorestframework==3.8.2
gevent==1.3.5
greenlet==0.4.14
gunicorn==19.9.0
Markdown==2.6.11
psycopg2-binary==2.7.5
pytz==2018.5
djangorestframework-jwt==1.11.0
coreapi==2.3.3
pygments==2.2.0
colorlog==3.1.4